chore: move form-contents component
