+++ src/components/form-contents/configured-form-contents.jsx
@@ -2,7 +2,7 @@ import { reduxForm } from 'redux-form';
 
 import FormContentsComponent from './form-contents.component';
 import validate from './form-contents-validations';
-import { asyncValidateDatasetInvokePatch } from '../formsLib/asyncValidateDatasetInvokePatch';
+import { asyncValidateDatasetInvokePatch } from '../../pages/dataset-registration-page/formsLib/asyncValidateDatasetInvokePatch';
 
 const config = {
   form: 'contents',
